// src/app/credit-request/models/loansModuleLoan.ts
/**
 * OMRO DigitalApp API
 * API used on OMRO Digital Resources
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { LoansModuleLoanAlert } from './loansModuleLoanAlert';
import { LoansModuleLoanMonthlyPaymentSummary } from './loansModuleLoanMonthlyPaymentSummary';

/**
 * Information added to DashboardModules array of UIOutputModel
 */
export interface LoansModuleLoan { 
    /**
     * Id of the loan as specified in the contract
     */
    loanId?: string;
    /**
     * Remaining unpaid amount
     */
    amountRemained?: number;
    /**
     * Currency ISO code (RON, EUR...)
     */
    currency?: string;
    /**
     * Curency display name (lei, euro ...)
     */
    currencyDisplayName?: string;
    /**
     * Total amount of the loan
     */
    totalAmount?: number;
    /**
     * The amount already paid by the customer
     */
    totalPaidAmount?: number;
    /**
     * The unpaid amount (equal to TotalAmount - TotalPaidAmount)
     */
    totalLeftAmount?: number;
    /**
     * Total overdue amount (overdue principal + interest + other )
     */
    totalOverdueAmount?: number;
    /**
     * Duration of the loan in months
     */
    months?: number;
    /**
     * Title of the loan
     */
    purpose?: string;
    /**
     * Ordered array of payment information (payment schedule)
     */
    payments?: Array<LoansModuleLoanMonthlyPaymentSummary>;
    /**
     * If not NULL, it contains the information to be displayed as a info card
     */
    alert?: LoansModuleLoanAlert;
    /**
     * The date of the next due payment date. NULL if there are no more payments
     */
    nextDueDate?: Date;
    /**
     * How many days until the next due date. NULL if there are no more payments
     */
    nextDueDateDays?: number;
    /**
     * Interest rate as a number between 0 and 100
     */
    interestRate?: number;
    /**
     * True if this loan is completely paid
     */
    isCompletelyPaid?: boolean;
    /**
     * Text describing the icon to be used when selecting the loan
     */
    iconHint?: string;
    /**
     * RGB color for the loan in CSS format #RRGGBB
     */
    color?: string;
    omroIban?: string;
    omroBank?: string;
}